Fix build issues with rollup
This one took me a while, but it turns out that Rollup requires Typescript to compile to ES Modules
before it can bundle and tree shake the application. Then, through the rollup/plugin-typescript,
you can specify the output format for rollup to transpile the ES Module Javascript too. This through me
for a loop because I needed the final output to be CommonJS and the formatting wasn't correct.